Fix 4d2362a0 CI change: Skip `no-commit-to-branch` `pre-commit` hook in GitHub Actions
The `no-commit-to-branch` hook was recently added (4d2362a0) to
`pre-commit` which fails if you attempt to commit directly to
`main`/`master`. This is a good check to run locally, but merging the PR
that added this hook into `main` caused CI to fail because it commits
directly to `main`. This commit skips `no-commit-to-branch` in GitHub
Actions. This fix was tested on `austin/fix_ci` and should fix CI once
merged into `main`.